refactor(virtual-mcp): extract ConnectionItem card into its own file
